SWITCHYARD-2474 Cleaned up dependencies and removed Annotations class
